tests done server done unittest can't run all at one (fix on its way) cli to do

+++ back/tasks.py
@@ -3,7 +3,7 @@ from back.auth import (
                         product_service_authentication_authorization
                       )
 from back.config import app, db
-from back.models import Application, Priority, Tasks, User
+from back.models import Application, Priority, Tasks, User, RequestStatus
 from back.utils import generate_uuid
 from flask import request, Response
 from sqlalchemy.exc import SQLAlchemyError
@@ -33,7 +33,7 @@ def create_task(*args):
         task_id=generate_uuid(),
         project_reference=project_id,
         description=decode_request.get("description"),
-        assign_to_user=user_assigned,
+        assign_to_user=decode_request.get("user_assigned"),
         priority=eval("Priority.{}".format(decode_request.get("priority").lower()))
     )
     try:
@@ -42,36 +42,52 @@ def create_task(*args):
         return Response(status=200)
     except SQLAlchemyError as e:
         return {
-            "error": e.__dict__["orig"]
+            "error": str(e.__dict__["orig"])
         }, 400
 
 
-@app.route("/retrieve_task/", methods=["GET", "DELETE"])
+@app.route("/retrieve_task/", methods=["GET", "PUT", "DELETE"])
 @product_service_authentication_authorization
-def retrieve_task(*args):
+def retrieve_task(user: User):
     decoded_request = json.loads(request.data.decode())
     if request.method == "GET":
-        if decoded_request.get("task_id"):
-            specific_task = Tasks.query.filter(
-                Tasks.task_id == decoded_request.get("task_id")
-            ).all()
-            if not specific_task:
-                return {
-                    "error": "No task found for this specific id"
-                }, 400
-            return specific_task, 200
         project_reference = decoded_request.get("project_reference")
         if not project_reference:
-            return {
-                "error": "Indicate project reference"
-            }, 400
+            specific_task = Tasks.query.filter(
+                Tasks.assign_to_user == user.username
+            ).all()
+            return {"tasks": [task.to_dict() for task in specific_task]}, 200
         tasks = Tasks.query.filter(
+            Tasks.assign_to_user == user.username,
             Tasks.project_reference == project_reference
         ).all()
         return {
             "project_reference": project_reference,
             "tasks": [task.to_dict() for task in tasks]
         }, 200
+    elif request.method == "PUT":
+        new_status = decoded_request.get("status")
+        task_concerned = decoded_request.get("task_id")
+        if not new_status or not task_concerned:
+            return {
+                "error": "Missing information (id or new status)"
+            }, 400
+        if new_status.lower() not in RequestStatus.__members__:
+            return {
+                       "error": "Status not recognized"
+                   }, 400
+        task_concerned = Tasks.query.filter(
+            Tasks.task_id == task_concerned
+        ).first()
+        task_concerned.status = eval("RequestStatus.{}".format(new_status))
+        try:
+            db.session.commit()
+            return Response(status=200)
+        except SQLAlchemyError as e:
+            return {
+                       "error": str(e.__dict__["orig"])
+                   }, 400
+        # either suspended or done or dismissed
     elif request.method == "DELETE":
         task_id = decoded_request.get("task_id")
         if not task_id:
@@ -79,13 +95,18 @@ def retrieve_task(*args):
                 "error": "No task id, required for deletion"
             }, 400
         task_to_delete = Tasks.query.filter(
-            Tasks.task_id == task_id
-        )
+            Tasks.task_id == task_id,
+            Tasks.status == RequestStatus.dismissed
+        ).first()
+        if not task_to_delete:
+            return {
+                "error": "id not recognized"
+            }, 401
         try:
             db.session.delete(task_to_delete)
             db.session.commit()
             return Response(status=200)
         except SQLAlchemyError as e:
             return {
-                "error": e.__dict__["orig"]
+                "error": str(e.__dict__["orig"])
             }, 400
